Extract common draw part from swing interop and Skia layer
